improved eks plugin design
